+++ crankshaft-console/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
+* Added handling for the task resource usage event ([#86](https://github.com/stjude-rust-labs/crankshaft/pull/86)).
 * Added cancellation of tasks from the TUI ([#53](https://github.com/stjude-rust-labs/crankshaft/pull/53))
 * Added initial TUI and server sync ([#46](https://github.com/stjude-rust-labs/crankshaft/pull/46)).
 * Added initial implementation ([#43](https://github.com/stjude-rust-labs/crankshaft/pull/43)).